moved out the airports to own file
